Reorganize Location related classes
