--- src/lib/Location.ts
@@ -1,4 +1,0 @@
-export class Location {
-	lat: number = 0;
-	lng: number = 0;
-}
