+++ src/lib/location.ts
@@ -6,16 +6,20 @@ export class Coordinates {
 		this.lng = lng;
 	}
 
-	lat: number;
-	lng: number;
+	lat!: number;
+	lng!: number;
 }
 
-export class Loc {
+export class CoordinatesWithLevel extends Coordinates {
+	level!: number;
+}
+
+export class Location {
 	constructor(coordinates: Coordinates, address: NewAddress){
 		this.coordinates = coordinates;
 		this.address = address;
 	}
 
-	coordinates: Coordinates;
-	address: NewAddress;
+	coordinates!: Coordinates;
+	address!: NewAddress;
 }
